--- src/main/java/dk/nykredit/jackson/dataformat/hal/annotation/Resource.java
@@ -1,14 +1,0 @@
-package dk.nykredit.jackson.dataformat.hal.annotation;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-/**
- * Annotation to mark something as being a HAL resource representation.
- */
-@Target({ElementType.TYPE })
-@Retention(RetentionPolicy.RUNTIME)
-public @interface Resource {
-}
