+++ src/main/resources/META-INF/services/com.fasterxml.jackson.core.ObjectCodec
@@ -1,1 +1,1 @@
-dk.nykredit.jackson.dataformat.hal.HALMapper
+io.openapitools.jackson.dataformat.hal.HALMapper
